Restructure TypeScript SDK into modular files
- Split tinybird.ts (2,087 lines) into 5 separate files by resource type:
  - datasources.ts: All 6 datasource definitions
  - pipes.ts: Internal pipe (analyticsHits)
  - materializations.ts: All 5 materialized view definitions
  - endpoints.ts: All 16 endpoint definitions
  - client.ts: createTinybirdClient export
- Update SDK version from 0.0.9 to 0.0.10
- Update tinybird.json to point to lib/client.ts
